move menus and context menus back to utils
