+++ src/editor/EditorMenu.js
@@ -1,6 +1,5 @@
-import Menu from "menus/Menu";
-
 import { ZOOMS } from "utils/CalchartUtils";
+import Menu from "utils/Menu";
 
 export default class EditorMenu extends Menu {
     getItems() {
